pyproject.toml
[project]
name = "google-sheets-cli"
version = "0.1.0"
description = "CLI tool for programmatic Google Sheets access with secure credential management"
authors = [
    {name = "Ebowwa Labs", email = "labs@ebowwa.com"}
]
dependencies = [
    "gspread",
    "google-auth-oauthlib",
    "requests",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []